refactor: Net*
