+++ springql-core/src/pipeline.rs
@@ -22,8 +22,8 @@ pub use name::{
     StreamName, ValueAlias,
 };
 pub use option::{
-    CANOptions, InMemoryQueueOptions, NetClientOptions, NetProtocol, NetServerOptions, Options,
-    OptionsBuilder,
+    CANOptions, InMemoryQueueOptions, NetProtocol, NetServerOptions, Options, OptionsBuilder,
+    SourceNetClientOptions,
 };
 pub use pipeline_graph::{Edge, PipelineGraph};
 pub use pipeline_version::PipelineVersion;
